Changes kernels to time based wait loops.
Changes NVTX ranges to get more information.
It turns out NVTX somtimes assigns the ranges in coroutines to the wrong thread. The range seemingly gets assigned as ending in the thread where the second part of the coroutine is executed, while redundant printouts show that the coroutine part (and the range) are fully running in the starting thread, as exepcted...