Reworks the architecture of the scheduler
There is now only a single queue, the work queue, for asynchronous
algorithm invocation. The queue usage is kept to a minimum and the
and scheduling of next algorithm or event is done right after
algorthm completion. Furthermore, the first next algorithm is
also invoqued immediately without queueing.

The event content manager allows detemining which algorithms
depend on the completed one and are ready to be executed.

Unit tests were added during the development.